initial tests

+++ ft_print_unsigned.c
@@ -1,10 +1,10 @@
 #include "printf.h"
 
-int	print_unsigned(unsigned int n)
+int	print_unsigned(unsigned long n)
 {
 	int	counter;
 	char	*str;
-
+	
 	counter = 0;
 	str = itoa_base(n, 10, "0123456789");
 	counter = sizenum_base(n, 10);
